test: split data loader tests by format
fix: change conftest to share fixtures rather than importing constants
